+++ functions.php
@@ -50,18 +50,31 @@ function genesis_sample_enqueue_scripts_styles() {
 	wp_enqueue_style( 'genesis-sample-fonts', '//fonts.googleapis.com/css?family=Source+Sans+Pro:400,600,700', array(), CHILD_THEME_VERSION );
 	wp_enqueue_style( 'dashicons' );
 
-	wp_enqueue_script( 'genesis-sample-responsive-menu', get_stylesheet_directory_uri() . '/js/responsive-menu.js', array( 'jquery' ), '1.0.0', true );
-	$output = array(
-		'mainMenu'          => __( 'Menu', 'genesis-sample' ),
-		'subMenu'           => __( 'Menu', 'genesis-sample' ),
-		'menuClasses'       => array(
-			'primary'   => '.nav-primary',
-			'secondary' => '.nav-secondary',
-		),
-		'menuIconClass'     => 'dashicons-before dashicons-menu',
-		'subMenuIconsClass' => 'dashicons-before dashicons-arrow-down-alt2',
-	);
+	$suffix = ( defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ) ? '' : '.min';
+	wp_enqueue_script( 'genesis-sample-responsive-menu', get_bloginfo( 'stylesheet_directory' ) . '/js/responsive-menus' . $suffix . '.js', array( 'jquery' ), CHILD_THEME_VERSION, true );
+	$output = base_responsive_menu_settings();
 	wp_localize_script( 'genesis-sample-responsive-menu', 'genesis_responsive_menu', $output );
+
+}
+
+// Define our responsive menu settings.
+function base_responsive_menu_settings() {
+	
+	$settings = array(
+		'mainMenu'          => __( 'Menu', 'genesis-sample' ),
+		'menuIconClass'     => 'dashicons-before dashicons-menu',
+		'subMenu'           => __( 'Menu', 'genesis-sample' ),
+		'subMenuIconsClass' => 'dashicons-before dashicons-arrow-down-alt2',
+		'menuClasses'       => array(
+			'combine' => array(
+				'.nav-primary',
+				'.nav-header',
+			),
+			'others'  => array(),
+		),
+	);
+
+	return $settings;
 
 }